Remove double domain from host/fqdn
It seems that DS appends the domain name to the host/fqdn from
lego. This results in e.g. _acme-challenge.example.com turning into
_acme-challenge.example.com.example.com and the challenge fails
because lego doesn't detect the key it was expecting. If the extra
domain is removed before creating the record, then it works. So
let's do that.

This issue may be connected to the dot-removal from a previous
commit. But it works fine like this.